+++ src/asr/asr_utils.py
@@ -147,43 +147,6 @@ class CompareValueTrigger(object):
         self._summary = chainer.reporter.DictSummary()
 
 
-def restore_snapshot(model, snapshot, load_fn=chainer.serializers.load_npz):
-    '''Extension to restore snapshot'''
-    @training.make_extension(trigger=(1, 'epoch'))
-    def restore_snapshot(trainer):
-        _restore_snapshot(model, snapshot, load_fn)
-
-    return restore_snapshot
-
-
-def _restore_snapshot(model, snapshot, load_fn=chainer.serializers.load_npz):
-    load_fn(snapshot, model)
-    logging.info('restored from ' + str(snapshot))
-
-
-def adadelta_eps_decay(eps_decay):
-    '''Extension to perform adadelta eps decay'''
-    @training.make_extension(trigger=(1, 'epoch'))
-    def adadelta_eps_decay(trainer):
-        _adadelta_eps_decay(trainer, eps_decay)
-
-    return adadelta_eps_decay
-
-
-def _adadelta_eps_decay(trainer, eps_decay):
-    optimizer = trainer.updater.get_optimizer('main')
-    # for chainer
-    if hasattr(optimizer, 'eps'):
-        current_eps = optimizer.eps
-        setattr(optimizer, 'eps', current_eps * eps_decay)
-        logging.info('adadelta eps decayed to ' + str(optimizer.eps))
-    # pytorch
-    else:
-        for p in optimizer.param_groups:
-            p["eps"] *= eps_decay
-            logging.info('adadelta eps decayed to ' + str(p["eps"]))
-
-
 class PlotAttentionReport(extension.Extension):
     """Plot attention reporter
 
@@ -241,6 +204,43 @@ class PlotAttentionReport(extension.Extension):
         plt.close()
 
 
+def restore_snapshot(model, snapshot, load_fn=chainer.serializers.load_npz):
+    '''Extension to restore snapshot'''
+    @training.make_extension(trigger=(1, 'epoch'))
+    def restore_snapshot(trainer):
+        _restore_snapshot(model, snapshot, load_fn)
+
+    return restore_snapshot
+
+
+def _restore_snapshot(model, snapshot, load_fn=chainer.serializers.load_npz):
+    load_fn(snapshot, model)
+    logging.info('restored from ' + str(snapshot))
+
+
+def adadelta_eps_decay(eps_decay):
+    '''Extension to perform adadelta eps decay'''
+    @training.make_extension(trigger=(1, 'epoch'))
+    def adadelta_eps_decay(trainer):
+        _adadelta_eps_decay(trainer, eps_decay)
+
+    return adadelta_eps_decay
+
+
+def _adadelta_eps_decay(trainer, eps_decay):
+    optimizer = trainer.updater.get_optimizer('main')
+    # for chainer
+    if hasattr(optimizer, 'eps'):
+        current_eps = optimizer.eps
+        setattr(optimizer, 'eps', current_eps * eps_decay)
+        logging.info('adadelta eps decayed to ' + str(optimizer.eps))
+    # pytorch
+    else:
+        for p in optimizer.param_groups:
+            p["eps"] *= eps_decay
+            logging.info('adadelta eps decayed to ' + str(p["eps"]))
+
+
 def torch_snapshot(savefun=torch.save,
                    filename='snapshot.ep.{.updater.epoch}'):
     """Returns a trainer extension to take snapshots of the trainer for pytorch."""
@@ -283,40 +283,6 @@ def _torch_snapshot_object(trainer, target, filename, savefun):
         shutil.move(tmppath, os.path.join(trainer.out, fn))
     finally:
         shutil.rmtree(tmpdir)
-
-
-def torch_resume(snapshot_path, trainer):
-    """Function to resume from snapshot for pytorch
-
-    :param str snapshot_path: snapshot file path
-    :param instance trainer: chainer trainer instance
-    :return: trainer instance whose states are loaded
-    """
-    # load snapshot
-    snapshot_dict = torch.load(snapshot_path, map_location=lambda storage, loc: storage)
-
-    # restore trainer states
-    d = NpzDeserializer(snapshot_dict['trainer'])
-    d.load(trainer)
-
-    # restore model states
-    if hasattr(trainer.updater.model, "model"):
-        # (for TTS model)
-        if hasattr(trainer.updater.model.model, "module"):
-            trainer.updater.model.model.module.load_state_dict(snapshot_dict['model'])
-        else:
-            trainer.updater.model.model.load_state_dict(snapshot_dict['model'])
-    else:
-        # (for ASR model)
-        if hasattr(trainer.updater.model, "module"):
-            trainer.updater.model.module.load_state_dict(snapshot_dict['model'])
-        else:
-            trainer.updater.model.load_state_dict(snapshot_dict['model'])
-
-    # retore optimizer states
-    trainer.updater.get_optimizer('main').load_state_dict(snapshot_dict['optimizer'])
-
-    return trainer
 
 
 # * -------------------- language model related -------------------- *
@@ -371,3 +337,61 @@ def get_model_conf(model_path, conf_path=None):
     with open(model_conf, "rb") as f:
         logging.info('reading a config file from ' + model_conf)
         return json.load(f, object_hook=AttributeDict)
+
+
+def torch_save(path, model):
+    """Function to save torch model states
+
+    :param str path: file path to be saved
+    :param torch.nn.Module model: torch model
+    """
+    if hasattr(model, 'module'):
+        torch.save(model.module.state_dict(), path)
+    else:
+        torch.save(model.state_dict(), path)
+
+
+def torch_load(path, model):
+    """Function to load torch model states
+
+    :param str path: file path to be loaded
+    :param torch.nn.Module model: torch model
+    """
+    if hasattr(model, 'module'):
+        model.module.load_state_dict(torch.load(path))
+    else:
+        model.load_state_dict(torch.load(path))
+
+
+def torch_resume(snapshot_path, trainer):
+    """Function to resume from snapshot for pytorch
+
+    :param str snapshot_path: snapshot file path
+    :param instance trainer: chainer trainer instance
+    :return: trainer instance whose states are loaded
+    """
+    # load snapshot
+    snapshot_dict = torch.load(snapshot_path, map_location=lambda storage, loc: storage)
+
+    # restore trainer states
+    d = NpzDeserializer(snapshot_dict['trainer'])
+    d.load(trainer)
+
+    # restore model states
+    if hasattr(trainer.updater.model, "model"):
+        # (for TTS model)
+        if hasattr(trainer.updater.model.model, "module"):
+            trainer.updater.model.model.module.load_state_dict(snapshot_dict['model'])
+        else:
+            trainer.updater.model.model.load_state_dict(snapshot_dict['model'])
+    else:
+        # (for ASR model)
+        if hasattr(trainer.updater.model, "module"):
+            trainer.updater.model.module.load_state_dict(snapshot_dict['model'])
+        else:
+            trainer.updater.model.load_state_dict(snapshot_dict['model'])
+
+    # retore optimizer states
+    trainer.updater.get_optimizer('main').load_state_dict(snapshot_dict['optimizer'])
+
+    return trainer
